wip: move some common functions to utils
